Fix file transfers
When the refactored BufferedMCLReader reaches the end of the segment
it keeps returning LineType.PROMPT until it is reset.

Usually this reset is done automatically by the BufferedMCLWriter.

However, in the file tranfer protocol there is a moment
where the server first acknowledges the end of the upload by sending
a segment containing PROMPT3, and then immediately another segment
containing the "&1 bla bla" status message.
Because the client doesn't send anything between these two, the
reader did not get reset.